moving utils to neu.tts.
