--- speech-synthesis/TTS/waveglow/utils/__init__.py
@@ -1,21 +1,0 @@
-# Copyright (c) 2017 Sony Corporation. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from pathlib import Path
-import sys
-
-from nnabla.ext_utils import get_extension_context
-
-sys.path.append(str(Path(__file__).parents[4] / 'utils'))
-from neu.comm import CommunicatorWrapper
