+++ speech-synthesis/WaveGlow/main.py
@@ -16,6 +16,11 @@ import argparse
 import os
 from pathlib import Path
 
+import sys
+sys.path.append(str(Path().cwd().parents[1] / 'utils'))
+from neu.comm import CommunicatorWrapper
+from neu.tts.optimizer import Optimizer
+
 import nnabla as nn
 from nnabla.ext_utils import get_extension_context
 from nnabla.logger import logger
@@ -27,9 +32,7 @@ from dataset import LJSpeechDataSource
 from hparams import hparams as hp
 from model.model import WaveGlow
 from train import WaveGlowTrainer
-from utils import CommunicatorWrapper
-from utils.optim import AnnealingScheduler
-from utils.optim import Optimizer
+from scheduler import AnnealingScheduler
 
 
 def run(args):
